api-demo/requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.0.1
amqp==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.3
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.1
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
cachetools==4.2.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.27.0
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
dulwich==0.21.5
dvc-http==2.30.2
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.0
fastjsonschema==2.18.0
ffmpy==0.3.1
filelock==3.12.0
flatten-dict==0.4.2
fonttools==4.41.1
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
funcy==2.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gpustat==1.1
gradio==3.38.0
gradio_client==0.2.10
grpcio==1.56.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
ipykernel==6.25.0
ipython==8.12.2
isoduration==20.11.0
iterative-telemetry==0.0.8
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kiwisolver==1.4.4
linkify-it-py==2.0.2
lit==16.0.6
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
nanotime==0.5.2
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
notebook_shim==0.2.3
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==11.525.112
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
orjson==3.9.2
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pipdeptree==2.9.3
pkgutil_resolve_name==1.3.10
platformdirs==3.9.1
prometheus-client==0.17.1
prompt-toolkit==3.0.38
protobuf==3.20.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
pydot==1.4.2
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rpds-py==0.9.2
rsa==4.9
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
safetensors==0.3.1
semantic-version==2.10.0
Send2Trash==1.8.2
shortuuid==1.0.11
shtab==1.6.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.23.1
uvloop==0.17.0
vine==5.0.0
voluptuous==0.13.1
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==3.15.0
